+++ pepper/rc/changelog.md
@@ -17,6 +17,7 @@
 - added `@output()` expansion which expands into an external command stdout
 - added support for recursive variable expansions (that is, it's possible to have a `@var(...)` inside a `@var(...)`)
 - changed parsing of expansion arguments to be more robust like command parsing
+- added `@command-source()` expansion which expands into a macro command's source
 
 # 0.30.0
 - added `insert-text` command